Added API to get hub and zappi serial numbers

+++ login-lambda/src/main/java/com/amcglynn/myzappi/login/rest/HubController.java
@@ -6,10 +6,13 @@ import com.amcglynn.myzappi.login.service.RegistrationService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
 
 /**
  * Handles the registering and retrieval of the myenergi hub/gateway.
  */
+@Slf4j
 public class HubController implements RestController {
 
     private final RegistrationService registrationService;
@@ -25,12 +28,26 @@ public class HubController implements RestController {
 
             return new Response(200, request.getBody());
         }
+        if (request.getMethod() == RequestMethod.GET) {
+            return get(request);
+        }
         if (request.getMethod() == RequestMethod.DELETE) {
             registrationService.delete(request.getUserId());
             return new Response(204);
         }
         System.out.println("Unsupported method for hub - " + request.getMethod());
         throw new ServerException(404);
+    }
+
+    @SneakyThrows
+    private Response get(Request request) {
+        var creds = registrationService.read(request.getSession().get().getUserId());
+        if (creds.isEmpty()) {
+            log.info("No hub found for user {}", request.getUserId());
+            return new Response(404);
+        }
+        var body = new ObjectMapper().writeValueAsString(creds.get());
+        return new Response(200, body);
     }
 
     public void register(Request request) {
